- Commented out functionalities that need to be inspected - Set default language to FI for development purposes - Fixed API URL on auth, set default language in App root, fixed unique key issue with child components - Deleted invalid component parameters - Fixed local auth to work without frontend proxy server - Refactored ISBN-ISMN form to improve functionality and readability - Fixed some logic errors in ISBN-ISMN form
