+++ src/components/form/constants.js
@@ -1,0 +1,13 @@
+export const PUBLICATION_TYPES = {
+	BOOK: 'book',
+	DISSERTATION: 'dissertation',
+	MAP: 'map',
+	MUSIC: 'music',
+	OTHER: 'other'
+};
+
+export const FORMATS = {
+	PRINT: 'printed',
+	ELECTRONIC: 'electronic',
+	BOTH: 'both'
+};
