+++ src/components/form/isbnIsmnRegistrationForm/constants.js
@@ -1,0 +1,12 @@
+export const PAGES = {
+	ADDITIONAL_DETAILS: 'additionalDetails',
+	AUTHOR_INFORMATION: 'authorInformation',
+	BASIC_INFORMATION: 'basicInformation',
+	CONTACT_INFORMATION: 'contactInfo',
+	FORMAT: 'formatBasic',
+	PUBLISHER_INFORMATION: 'publisherBasicInfo',
+	PUBLISHING_ACTIVITIES: 'publishingActivities',
+	SERIES_INFORMATION: 'seriesInformation',
+	AVAILABILITY_INFORMATION: 'availabilityInformation',
+	UNIVERSITY_INFORMATION: 'universityInformation'
+};
