+++ src/components/form/isbnIsmnRegistrationForm/validation.js
@@ -1,50 +1,4 @@
-/**
- *
- * @licstart  The following is the entire license notice for the JavaScript code in this file.
- *
- * UI microservice of Identifier Services
- *
- * Copyright (C) 2021 University Of Helsinki (The National Library Of Finland)
- *
- * This file is part of identifier-services-ui
- *
- * identifier-services-ui program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as
- * published by the Free Software Foundation, either version 3 of the
- * License, or (at your option) any later version.
- *
- * identifier-services-ui is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- *
- * @licend  The above is the entire license notice
- * for the JavaScript code in this file.
- *
- */
-
-export function readCookie(name) {
-	const nameEQ = `${name}=`;
-	const ca = document.cookie.split(';');
-	for (let i = 0; i < ca.length; i++) {
-		let c = ca[i];
-		while (c.charAt(0) === ' ') {
-			c = c.substring(1, c.length);
-		}
-
-		if (c.indexOf(nameEQ) === 0) {
-			return c.substring(nameEQ.length, c.length);
-		}
-	}
-
-	return null;
-}
-
 /* eslint-disable complexity */
-
 export function validate(values) {
 	const errors = {
 		publicationDetails: {frequency: {}},
@@ -54,7 +8,7 @@ export function validate(values) {
 		university: {}
 	};
 	const {
-		publicationDetails = {frequency: {}},
+		publicationDetails = {},
 		previousPublication = {},
 		postalAddress = {},
 		formatDetails = {},
@@ -88,11 +42,13 @@ export function validate(values) {
 		'langGroup',
 		'category',
 		'rangeStart',
-		'rangeEnd'
+		'rangeEnd',
+		'phone'
 	];
+
 	requiredFields.forEach(field => {
 		if (!values[field]) {
-			errors[field] = 'Required';
+			errors[field] = 'field.required';
 		}
 	});
 
@@ -121,46 +77,46 @@ export function validate(values) {
 	}
 
 	if (!values.selectUniversity) {
-		errors.selectUniversity = 'Required';
+		errors.selectUniversity = 'field.required';
 	}
 
 	if (!university.name) {
-		errors.university.name = 'Required';
+		errors.university.name = 'field.required';
 	}
 
 	if (!university.city) {
-		errors.university.city = 'Required';
+		errors.university.city = 'field.required';
 	}
 
 	if (!values.contactEmail) {
-		errors.contactEmail = 'Required';
+		errors.contactEmail = 'field.required';
 	} else if (!/^[A-Z0-9._%+-]+@[A-Z0-9.-]+\.[A-Z]{2,4}$/i.test(values.contactEmail)) {
-		errors.contactEmail = 'Invalid e-mail address';
+		errors.contactEmail = 'format.email';
 	}
 
 	if (!formatDetails.format) {
-		errors.formatDetails.format = 'Required';
+		errors.formatDetails.format = 'field.required';
 	} else if (formatDetails.format.value === '') {
-		errors.formatDetails.format = 'Required';
+		errors.formatDetails.format = 'field.required';
 	}
 
 	if (!formatDetails.fileFormat) {
 		if (!formatDetails.otherFileFormat) {
-			errors.formatDetails.fileFormat = 'Required';
+			errors.formatDetails.fileFormat = 'field.required';
 		}
 	} else if (formatDetails.fileFormat.value === '') {
 		if (!formatDetails.otherFileFormat) {
-			errors.formatDetails.fileFormat = 'Required';
+			errors.formatDetails.fileFormat = 'field.required';
 		}
 	}
 
 	if (!formatDetails.printFormat) {
 		if (!formatDetails.otherPrintFormat) {
-			errors.formatDetails.printFormat = 'Required';
+			errors.formatDetails.printFormat = 'field.required';
 		}
 	} else if (formatDetails.printFormat.value === '') {
 		if (!formatDetails.otherPrintFormat) {
-			errors.formatDetails.printFormat = 'Required';
+			errors.formatDetails.printFormat = 'field.required';
 		}
 	}
 
@@ -168,61 +124,64 @@ export function validate(values) {
 		errors.formatDetails.run = 'Must be a number [0-9]';
 	}
 
-	if (publicationDetails.frequency) {
-		if (!publicationDetails.frequency.currentYear) {
-			errors.publicationDetails.frequency.currentYear = 'Required';
-		} else if (!/^([0-9]|([MDCLXVI]))*$/gm.test(publicationDetails.frequency.currentYear)) {
-			errors.publicationDetails.frequency.currentYear = 'Must be a number [0-9] or Roman';
-		}
+	if (!publicationDetails.currentYearFrequency) {
+		errors.publicationDetails.currentYearFrequency = 'field.required';
+	} else if (!/^([0-9]|([MDCLXVI]))*$/gm.test(publicationDetails.currentYearFrequency)) {
+		errors.publicationDetails.currentYearFrequency = 'format.integer';
+	}
 
-		if (!publicationDetails.frequency.nextYear) {
-			errors.publicationDetails.frequency.nextYear = 'Required';
-		} else if (!/^([0-9]|([MDCLXVI]))*$/gm.test(publicationDetails.frequency.nextYear)) {
-			errors.publicationDetails.frequency.nextYear = 'Must be a number [0-9] or Roman';
-		}
-	} else {
-		errors.publicationDetails.frequency.nextYear = 'Required';
-		errors.publicationDetails.frequency.currentYear = 'Required';
+	if (!publicationDetails.nextYearFrequency) {
+		errors.publicationDetails.nextYearFrequency = 'field.required';
+	} else if (!/^([0-9]|([MDCLXVI]))*$/gm.test(publicationDetails.nextYearFrequency)) {
+		errors.publicationDetails.nextYearFrequency = 'format.integer';
 	}
 
 	if (previousPublication.lastYear && !/^([0-9]|([MDCLXVI]))*$/gm.test(previousPublication.lastYear)) {
-		errors.previousPublication.lastYear = 'Must be a number [0-9] or Roman';
+		errors.previousPublication.lastYear = 'format.integer';
 	}
 
 	if (previousPublication.lastNumber && !/^([0-9]|([MDCLXVI]))*$/gm.test(previousPublication.lastNumber)) {
-		errors.previousPublication.lastNumber = 'Must be a number [0-9] or Roman';
+		errors.previousPublication.lastNumber = 'format.integer';
 	}
 
 	if (!publicationDetails.previouslyPublished) {
-		errors.publicationDetails.previouslyPublished = 'Required';
+		errors.publicationDetails.previouslyPublished = 'field.required';
 	}
 
 	if (!publicationDetails.publishingActivities) {
-		errors.publicationDetails.publishingActivities = 'Required';
+		errors.publicationDetails.publishingActivities = 'field.required';
 	}
 
 	if (isNaN(Date.parse(values.publicationTime))) {
-		errors.publicationTime = 'Not Valid date';
+		errors.publicationTime = 'format.date';
 	}
 
 	if (!postalAddress.address) {
-		errors.postalAddress.address = 'Required';
+		errors.postalAddress.address = 'field.required';
 	}
 
 	if (!postalAddress.city) {
-		errors.postalAddress.city = 'Required';
+		errors.postalAddress.city = 'field.required';
 	}
 
 	if (!postalAddress.zip) {
-		errors.postalAddress.zip = 'Required';
+		errors.postalAddress.zip = 'field.required';
 	}
 
-	// If (!/^[a-zA-ZÀ-ÿ\s@~`!@#$%^&*()_=+\\';:"\/?>.<,-]{3,256}$/i.test(values.name)) {
-	// 	errors.name = 'Name should contains only 3-256 alphabets';
-	// }
+	if (postalAddress.zip && !/^\d+$/.test(postalAddress.zip)) {
+		errors.postalAddress.zip = 'postalAddress.zip.format';
+	}
+
+	if (postalAddress.city && !/^[a-zA-Z\s]+$/.test(postalAddress.city)) {
+		errors.postalAddress.city = 'postalAddress.city.format';
+	}
+
+	if (values.phone && !/^[0-9-+]+/.test(values.phone)) {
+		errors.phone = 'format.phone';
+	}
 
 	if (!/^[A-Z0-9._%+-]+@[A-Z0-9.-]+\.[A-Z]{2,4}$/i.test(values.publisherEmail)) {
-		errors.publisherEmail = 'Invalid e-mail address';
+		errors.publisherEmail = 'format.email';
 	}
 
 	if (values.primaryContact && values.primaryContact.length > 0) {// Empty
@@ -235,22 +194,22 @@ export function validate(values) {
 
 	function validateContact() {
 		if (!values.email) {
-			errors.email = 'Required';
+			errors.email = 'field.required';
 		} else if (!/^[A-Z0-9._%+-]+@[A-Z0-9.-]+\.[A-Z]{2,4}$/i.test(values.email)) {
-			errors.email = 'Invalid e-mail address';
+			errors.email = 'format.email';
 		}
 	}
 
 	if (!/\w{2,}/i.test(values.streetAddress)) {
-		errors.streetAddress = 'Value must be between more than 2 characters';
+		errors.streetAddress = 'format.length';
 	}
 
 	if (!/\w{2,}/i.test(values.city)) {
-		errors.city = 'Value must be between more than 2 characters';
+		errors.city = 'format.length';
 	}
 
 	if (!/^\d{3,}$/i.test(values.zip)) {
-		errors.zip = 'Value must be numbers';
+		errors.zip = 'format.integer';
 	}
 
 	if (values.authors && values.authors.length > 0) { // Empty
@@ -266,11 +225,11 @@ export function validate(values) {
 
 	function validateAuthor() {
 		if (!values.authorGivenName) {
-			errors.authorGivenName = 'Required';
+			errors.authorGivenName = 'field.required';
 		} else if (!values.authorFamilyName) {
-			errors.authorFamilyName = 'Required';
+			errors.authorFamilyName = 'field.required';
 		} else if (!values.role) {
-			errors.role = 'Required';
+			errors.role = 'field.required';
 		}
 	}
 
@@ -283,34 +242,31 @@ export function validate(values) {
 	};
 
 	if (!values.classification) {
-		errors.classification = 'Required';
+		errors.classification = 'field.required';
 	}
 
 	if (!values._id && !frequency.value) {
-		errors.frequency = 'Required';
+		errors.frequency = 'field.required';
 	}
 
 	if (!values._id && !type.value) {
-		errors.type = 'Required';
+		errors.type = 'field.required';
 	}
 
 	if (issnFormatDetails === null || issnFormatDetails.length === 0) {
-		errors.issnFormatDetails = 'Required';
+		errors.issnFormatDetails = 'field.required';
 	} else if (issnFormatDetails.length > 0) {
 		if (issnFormatDetails.some(item => item.value === 'online') && !formatDetails.url) {
-			errors.formatDetails.url = 'Required';
+			errors.formatDetails.url = 'field.required';
 		}
 	}
-
-	// ISBN/ISMN Registration form
-	// Availability information
 
 	if (!values.isPublic || String(values.isPublic).toLowerCase() !== 'true') {
 		errors.isPublic = 'publicationRegistrationIsbnIsmn.form.availability';
 	}
 
 	if (!values.publicationType) {
-		errors.publicationType = 'Required';
+		errors.publicationType = 'field.required';
 	}
 
 	return errors;
