+++ src/configuration.js
@@ -4,7 +4,7 @@
  *
  * UI microservice of Identifier Services
  *
- * Copyright (C) 2021 University Of Helsinki (The National Library Of Finland)
+ * Copyright (C) 2019 University Of Helsinki (The National Library Of Finland)
  *
  * This file is part of identifier-services-ui
  *
