+++ src/intl/getMessages.js
@@ -1,12 +1,4 @@
-import * as enMessages from './translations/en.json';
-import * as fiMessages from './translations/fi.json';
-import * as svMessages from './translations/sv.json';
-
-const translations = {
-	en: enMessages,
-	fi: fiMessages,
-	sv: svMessages
-};
+import {translations} from '../intl/translations';
 
 export function getMessages(locale) {
 	const messages = translations[locale];
